createdBlips table
modified the loop that creates the blips to first check if a blip with the same coordinates already exists in the createdBlips table. If it does, the loop skips that blip and moves on to the next one. If it doesn't, the blip is created as before, and its coordinates are added to the createdBlips table so that we know not to create another blip with the same coordinate

Updated Config.lua to suit